// frontend/src/types/course.types.ts
export interface CourseCardData{
id?:number;
slug:string;
title:string;
subTitle?: string;
thumbnail:string;
difficulty:string;
duration:number;
instructorName:string;
isInWishList:boolean;
isEnrolled?: boolean;
}